move: import에 변경된 파일 위치 반영
