+++ views/javascript/components/outputPopup.js
@@ -1,5 +1,5 @@
 import { DomManager } from '../utils/domManager.js';
-import { BasePopup } from './basePopup.js';
+import { BasePopup } from './base/basePopup.js';
 
 export class OutputPopup extends BasePopup {
   constructor(title, content, applyCallback) {
@@ -8,6 +8,15 @@ export class OutputPopup extends BasePopup {
       document.getElementById('overlay'),
     );
     this.set(title, content, applyCallback, () => this.hide());
+  }
+
+  show(title, content, okCallback, cancelCallback = null) {
+    super.show(title, content, okCallback, cancelCallback);
+
+    const radioBtnGroup = this.holder.querySelector('.radio-btn-group');
+    if (radioBtnGroup) {
+      DomManager.hideElement(radioBtnGroup);
+    }
   }
 
   hideButton() {
