+++ views/javascript/components/spinner.js
@@ -1,4 +1,4 @@
-import { BaseComponent } from './baseComponent.js';
+import { BaseComponent } from './base/baseComponent.js';
 
 export class Spinner extends BaseComponent {
   constructor(holder) {
